move to trait
